requirements.txt
torch
schedulefree
tensorboard
pydantic
lxml
hsluv
nevergrad
requests
pillow
